+++ doc/RELEASE_NOTES.md
@@ -3,3 +3,8 @@
 - Demo
 
 ## v0.1.0-build-2
+
+- Added transition Animation for People profile. New transition animation is available in
+  androidx.compose.ui since v1.7.0 alpha and navigation-compose 2.8.0 alpha.
+
+## v0.1.0-build-3